+++ Models/Project.cs
@@ -4,12 +4,14 @@ namespace Review.Api.Models;
 
 public class Project : BaseEntity
 {
-    public Project(string name)
+    public Project(string name, string createdByUserId)
     {
         Name = name;
+        CreatedByUserId = createdByUserId;
     }
 
     public string Name { get; set; }
+    public string CreatedByUserId { get; set; }
 
     public List<User> Users { get; } = [];