+++ Services/IProjectService.cs
@@ -1,4 +1,3 @@
-using Review.Api.Models;
 using Review.Api.Models.DTOs;
 
 namespace Review.Api.Services;